move floor function to math module
